Fix problem of first scan without any point
First scans after the Pool restart, were not able to report
any scan points. This was due to missing event configuration
of the Data attribute.

Configure CHANGE event without detection for the Data attr.

Implement read_Data so it raises the DevFailed exception with the
"UnsupportedFeature" reason.

Improve the OnDataChangedCb listener:
* ignore "UnsupportedFeature" exception
* raise any other exceptions

refs #5136