Implement sardana event generation in DummyTGCtrl
Add RectangularFuncionGenerator class which implements the necessary logic
of the DummyTriggerGateController: generation of the sequence of events
simulating the periodic rectangular function.

Implement AddDevice, StartOne, AbortOne and State one so the generation
of events is possible.

Implement SetAxisParam and GetAxisParam for configuration of the following
parameters: offset, active_period, passive_period and repetitions.