Encapsulate value and index in lists
At this moment experimental channel values are passed via two
different Tango attributes (Value and Data).
The discrimination is  based on type of the value: if value's type
is scalar it is passed via Value, and if its type is spectrum it
is passed via Data. We want to pass it via Data.
Encapsulate value and index in lists.