Refactor AscanctTest to use a on the fly Sardana
AscanctTest inherits from the different MeasSarTestTestCase
and BaseMacroServerTestCase helpers that are in charge of create
a pool, the elements (like motors, exp. channels, etc.)
and macroserver.

Allowing test a macro in a clean and sole environment for this
test. At the end of the test, the system will be deleted.