Fix already involved in operation exceptions
Weak condition of only checking the action's state causes
aleatory already in operation exceptions. This is due to the fact
that more than one action's execution could be put in the Worker's queue.
Allow to execute software synchronized acquisition, only if there is none
being executed neither waiting in the queue.