Fix bug-121: Add macro libraries if they have macros
Add macro libraries in Sardana only if they contain macros.
If the libraries does not contain any macro, they will not be listed
with the spock command 'lsmaclib'.

If the macro library module contain non valid macros, the
module will not be listed neither. The libraries will not be present
in the Tango attribute 'MacroLibList'.