Fix bug-256: do not look for libraries inside MacroPath
Libraries used by macros (but not containing any macro) have to be
located in directories included in the system pythonpath
or in directories defined in the property PythonPath from
the MacroServer Tango Device.

These libraries do not have to be located in the MacroPath.
Only libraries containing macros (macro libraries), shall be located
in directories listed in the MacroPath.

Do not look for libraries not containing macros in directories not
included in the PythonPath.