Fix bug-238: Workaround position's w_value of range
Do not update the w_value if it is out of range: when updating position's
w_value simply ignore errors of API_WAttrOutsideLimit reason.
The position's w_value will not be updated (only when out of range) during
readouts or when pushing events.

The workaround does not affect the drift correction feature of the pseudomotors
but affects the pending operation feature of the Taurus write widgets of
the position attribute or any other feature (not known at the moment of
applying this workaround) or anyone trusting the w_value.

Remove this protection whenever feature-286 has been implemented and bug-54 has
been fixed. The lack of the feature and the bug can lead to the situation when
motor's position write value is out of range.

feature-286: "Solve inconsistencies between user position limits and dial
position limits" (http://sourceforge.net/p/sardana/tickets/286)

bug-54: "Software limits problems between motors and pseudomotors"
(http://sourceforge.net/p/sardana/tickets/54)