Modify timeouts in tests
Modify timeouts in macros tests in order to allow the execution
of the whole test suite, even in the event of tests that do not
finish in a certain time. If some of the tests elapses more time
than the specified timeout, the tests fails and the test suite
continues its execution.