Add synchronization column in expconf
Add a new column to the expconf GUI for synchronization element.
Implement the logic to allow set a trigger/gate element for the
experimental channels.

As SEP6 says each controller only can have a trigger element. So, if any
of the channels of the same controller (involved in the current measurement group)
changes its syncrhonization element, this  will be modified for all involved channels
(of the same  controller) of the current measurement group.