Add hooks in TriggerGateControllerTestCase
Add empty post_configuration_hook and post_generation_hook methods
in the TriggerGateControllerTestCase class. These hooks are calling in
generation method (post calling PreStartOne and once the StartOne is finished)
and allow to do extra actions.

Modify the parametrization of the pregeneration method. Now it uses a
dictionary instead of several parameters.