Simplify synchronization to avoid Source and Value
SynchSource and SynchValue seems to be too generic. No application is known
to require to change source in the Position or Monitor domain in the same
measurement. Avoid using of them and pass the value directly as a float.
Both sources: position and monitor will be configured directly on the
measurement level.

Adopt tests to do the same.