Add getters of meta classes
Add getRecorderMetaClass and getRecorderMetaClasses.

Fix implementation of getRecorderClass and getRecorderClasses.

Change filter from BaseFileRecorder to DataRecorder/

At this moment tests with FakeScanRecorder do not run because of
SardanaModuleManager issue: test in macroserver package pretend to be
macroserver subpackage but after installation it is not (what confuse
loadModule function). LoadModule searches for sardana.macroserver.test in
installed version of sardana which does not exists. Solution could be to keep
tests separately from sardana.