Only consider position updates after start
Two consecutive scans using SoftwareTriggerGate in the position domain do not
work properly. This is due to the fact that the position listener is added
prior to the start of generation. When the position listener is added first
event arrives, and its value may break the generation of events.

Discard these values by resetting the event in the startup. So only updates
after the starup are considered as valid.