Feature-390
Macro relmac and relmaclib remove the library and its macros
if the new library has any error.
Make relmac and relmaclib loading the new macro library only
in case the new library is free of errors. Otherwise the old
library and their macros are kept.

- Add to sardanamodulemanager a method, isValidModule,
to check if a python module is loadable.

- Adapt reloadMacroLib method in msmacromanager to validate
the given macro library, and in case of error return an empty
library with errors description.

- Adapt the  MacroServer reload_macro_lib method to protect it
against empty macro libraries. Make reload_macro method return
the macro library.

- Minor changes in the relmac and relmaclib macros.