Fix bug-6: Correct PoolPath precedence
ControllerManager is not loading controller classes and modules correctly.
In case of overriding controller classes or modules, the lower priority items
are the ones taking precedence. This is due to the fact that only module names
are considered when discovering.

Consider a pair of module name and its full path while discovering and load
them in the reversed order of priorities, so the highest priority items may
override the lower priority items (both modules and classes) during the loading
process.