Fix info messages about ScanRecorder env variable
Info message "ScanRecorder is not defined. This operation will use the default
recorder." sounds like the error message (not info) and is not fully certain
by indicating that the default recorder will be used (SPEC recorder).
In this case a recorder from the recorder_map, i.e. related to the scan file
extension will be used.

Remove this message and add additional information when ScanRecorder environment
is in use.