Remove DataFormats enumerator in favor of strings
DataFormats values are used only as keys of formats dictionary.
For the user it may cause misleading information that only the defined file
extensions are supported by Sardana. Also according with the new definitions of
recorders (based on plugins) it does not have sense to maintain the enumeration
without offering the mechanism to add new formats to it.
Use simple strings instead.