Fix-bug431
It is a bug introduced by Feature-427.
The Door State does not recover from the state ALARM when a macro is
aborted for an error in the macro itself.

Adapt the is_RunMacro_allowed and is_GetMacroEnv_allowed methods in the
Door class to accept the Macro.Exception as a valid state.