Fix bug-451: avoid reset of params if empty str used
Macroexecutor's parameditor resets all values to None when one of params was
left (focus lose or apply) with an empty string. This is due to a buggy
validation of spock's line edit text.

Workaround it by: setting the parameter's default value (if present) or ignoring
the change to an empty string.