Fix bug-464: correct min&max in wa, pwa, ls*
pwa, wa, ls* are defined with the repeat parameter - filter.
The repeat parameter defines: min=0 and the default value .*.
The ParamDecoder does not create the extra repetition if min=0.

Change repeat parameter definition to specify the min=1, since the macro logic
requires it.

Change repeat parameter definition to work without limiting the max=1 - it works
perfectly.