Change API of ParamDecoder to use only params_def
ParamDecoder is using the meta macro object what but it is too much for its
needs. Simple params_def (from the macro definition) is enough.

This simplification makes the ParamDecoder class more independent to the
macroserver infrastructure (the need of creating the MacroManager) and allows
easier unittesting.

Change the API to receive just the params_def.