Add generic function to parse str params to xml
Multiple places in the sardana code require parsing string macro parameters
to the correct xml structure. The already existing places are: downstream flow
of the RunMacro Tango command and spock's macro execution. In the future
the execMacro macro API method will also require it.

Implement a generic function to provide this functionality in order to avoid
duplication of code.