Fix bug when loading nested param repeats
History and Favorite editors have problems with loading macros with nested
repeat parameters. The loaded structure of repeat nodes is mixed and erroneous.
This is because lxml.etree._Element iterator of a given tag is iterating over
nested tag elements as well.

Do not use tagged iterator and iterate just over the first level repeat nodes.