Fix bugs in sequencer due to value = None
ParamNodes values may stay as None (if not edited), this cause problems when
operating on strings e.g. concatenation or setting value to widgets e.g.
cast values to string to always show something, even if the parameter value
is not set yet (is None). This makes the parameditor more intuitive showing to
the user where to edit the parametrs values.