Consider ZeroD when evaluating MG's state
The ZeroD experimental channels are treated differently when the measurement
group's state is evaluated (they do not participate) and when the measurement
group's involvment in operation is evaluated (they participate). This causes
random errors in step and continuous scans - "already involved in operation".

Treat ZeroD equally to other experimental channels when evaluating the
measurement group's state.