Avoid frozen scans when ZeroD's readout fails
The uncaught exception (see below) in the 0D acquisition thread prevents the
change of ZeroD's state to ON. As a result, the MeasurementGroup never changes
state to ON what "frozes" the scan.

The exception is due to the fact that the Accumulation classes are not handing
well appending of None's (None indicate the ZeroD readout failure).
Readout failure (None) and it's timestamp are correctly appended to the buffers.
Fix accumulation calculations by introducing the following changes:
* SumAccumulation does ignores the failed readouts
* AverageAccumulation does the average only over the successful readouts

Be aware that the ZeroD's readout failures may still produce scan records with
<nodata> of ZeroD's [#418].