Implement optional data interpolation
The zero order interpolation is applied by default in the recordlist of a scan.

Make it optional via the definition of the boolean "ApplyInterpolation"
sardana environment variable. The GScan is in charge of configure the ScanDate
using this variable.

If this variable is not defined by default the zero order interpolation is
disabled.