Respect case in expconf meas. groups combobox
Measurement Group combobox is showing measurement group names with lower case.
Even if the Sardana element names are case insensitive, the user may still
introduce the names with different cases and the client should show them.
Extract the case information from the label field and show it.

Make finding the ActiveMntGrp in the combobox case insensitive since the
measurement groups are case insensitive.