Fix bug-508: make spock compatibile with IPython 5
Magic commands e.g. macros are not loaded into the spock session when using
IPython 5 (actually tested with 5.1). define_magic function, used by spock
to register the magics, was deprecated since IPython 1.0 and finally it was
removed. Use register_magic_function instead which is compatible along all
versions >= 1.

register_magic_function, when used with the line magics, expects functions with
just one parameter - the string line of the magic's arguments. Change
definition of the macro_fn to follow the API.