Fix bug-509 - add line break when dooroutputs are multi-lines
The DoorOutput widget does not break lines and prints them in continuation
of a signle line. Since we are appending the HTML text to the widget use
<br> tag instead of "\n" character to fix it.
Apart of that add the line break only between the lines and not after the
last one since every HTML append creates a new paragraph.