Treat ZeroD as slave of software acquisition
There are problems with the state machine of the measurement group.
Software synchronized elements may still not be started when the software
trigger gate already finished generation. In this case the measurement group's
state is considered as ON. This may lead to the "already involved in operation"
errors during the step scan.

Return to the solution of making the ZeroD acquisition slave of the SoftwareAcquisition.
Also start both actions in a synchronous way.