Fix bug-518: waitFinish is not reservedOperation
reservedOperation is not executed in the macro run method after the macro
was stopped e.g. in the finally clause when the stop occurred in the try
clause. In case of executing an asynchronous operation e.g. move or count
this forces to synchronize it i.e. call the waitFinish in the on_stop/on_abort
method. Otherwise the macroserver gets hung.

Don't threat the waitFinish method as the reservedOperation.