Improve handling of errors when reading pseudocounters
Pseudocounter errors are only based on errors of the underlying attributes.
This is not sufficient. Error may occur when reading an underlying attribute
(when it doesn't have a value yet) or when executing the calculation. Store
this errors in between calculations so the pseudocounter's value error can
be evaluated correctly.