Does not allow to update 0D value from outside of acquisition
Each Value attribute class must have an update method. In case of the 0D
it is not allowed to update it's value from outside of the 0D acquisition thread.
Otherwise it will require proper protection of the critical section.

Let's imagine an acquisition involving a pseudocounter based on a counter
and a 0D channel. If a counter updates its value, the pseudocounter will
require the 0D value to perform the calculation. Hence an update of the 0D
will be requested. Simultaneously the 0D acquisition thread will fill the
accumulation buffer. Without proper protection of the critical section this
won't be safe.