Adapt sardanaeditor to be compatible with Taurus4
The sardanaeditor uses Taurus signals that have been changed to new style
in Taurus 4. The connect method does not work anymore if the used signals
are passed as shortcut, when the signals expect input values.

Fix it, adding a back. comp. layer using new style in a try/except block.