Add SEPs access in Sardana documentation
SEPs are moved from SF wiki to GitHub. Sardana documentation
shall point to the place where the SEPs are hosted. Add
link to docs.rst page so that it is shown in the documentation
index.

Implement a redirect system to access to the SEPs from the
Sardana web page with the following syntax:

www.sardana-controls.org/sep/?SEPXX.EXT

where:
- XX is the number of the SEP. e.g. 1,2...
- EXT is the extension of the SEP document. e.g. md,...