Obsolete many SEPs according to SEP15
SEP15 obsoletes SEP0, SEP1, SEP7 and SEP10.
SEP1 and SEP10 references to SourceForge, and they became obsolete given
that the Sardana project moves to GitHub.

SEP0 and SEP7 describe the contribution process which has changed for
one which privileges de Pull Requests. SEPs are not located anymore
in a SourceForge wiki; now they are independent .md files of the
Sardana project.