Fix #386: equalize behavior of ct and scans if data is missing
I've realised that 'ct' macro behaves different than 'ascan' when an
element in the measurement group didn't bring back any data (like an
attribute that returns an exception when its read).

I've seen how the "scan" infrastructure does it to place in the cell of
the recorder to write a "<nodata>" string, and this patch tries to do
the most similar behaviour.