Do not export DialPosition and Position memorized values
Skip memorized values of DialPosition and Position. DialPosition because
it is read only attribute and the current version of sardana script
would not be able to set it. Position is no more a memorized attribute,
but the database may still contain the __value property for it - from the
previous versions of Sardana (Pool in C++).