TaurusMacroConfigurationDialog shows the MS disconected
TaurusMacroConfigurationDialog does not work if the default_tango_host
was set in a taurusgui.

It is due to the uses of PyTango, where the full_name does not include
the database information and it asumes the value of $TANGO_HOST env. var.

Fix it, composing the model with the database and attribute name.