Improve main in macrobutton.py module
The code enables to execute any macro (except param repeat
with mutiple values) from a macrobutton. Two arguments
are required to initalize the widget: door name and macro name.
Once the widget is running, the macro name can be changed and
the interface to supply the parameters is updated automatically.