Make 0D an independent action.
0D acquisition is slave to the CT acquisition. This fails in case the CT
acquisition fails on the startup - the 0D is never stopped. As a consequence
the "already involved in operation" error is raised when the measurement
group is started again. Change the approach and make these two actions
independent. The 0D acquisition is stopped by the passive even from the software
synchronizer.