Workaround for synchronization problems with acquisition
Acquisition and software synchronizer have synchronization problems.
This is because the acquisition actions are executed by the worker threads
and the synchronizer may already finish before the job actually starts.
This causes inconsistency with the measurement group elements' states.
Workaround it by using a started flag. In the future this should be improved
by rethinking the synchronization between the threads involved in the
acquisition.