Check repeat nodes names
Update test macro by checking the names of all RepeatNodes in
the given ParamRepeatNode. The RepeatNode names must be
correlative.

This is important because when duplicating a node,
we could be duplicating the name of the node that has to be
duplicated. It could happen as well, if duplication method
not well implemented, that the new node overwrites the name
of the node that has to be duplicated.