Base BufferedAttribute on a dedicated Buffer class
Implement a Buffer class with the possibility to access its elements via
index and maintaining the order of elements.

Enhance BufferAttribute class for the needs to handle pseudocounters.
set_value_chunk was splitted in two different methods: append_value_buffer
and extend_value_buffer.

This commit breaks the functionalities in the experimental channels.