Temporarily clear value_buffer at the beginning of acquisition.
For the moment we can not clear value buffers at the end of
the acquisition. This is because of the pseudo counters that are
based on channels synchronized by hardware and software.
These two acquisition actions finish at different moment so the
pseudo counter will loose the value buffer of some of its physicals
if we clear the buffer at the end. Clear it at the beginig of acquisition.