Fix dependency version specifiers
Some older versionos of setuptools does not recognize version specifiers
within parentheses. Change it to a format compatible with all versions.

Remove conditional dependency of ordereddict becasuse it is only
available in the very new versions of setuptools
(see: https://hynek.me/articles/conditional-python-dependencies/).
This means that ordereddict won't be installed with pip for Python 2.6.