Make compatibilities between Taurus3 and Taurus4
Sardana code is not fully ready to deal with Taurus4 model names.
To be compatible with Taurus4 the taurus models must be created with
full name containing the scheme (tango://). But all the recording
implelementations and the the measurement group configuration which
affects the results of the count method is still based on Taurus3 full
name models - the ones that do not contain the scheme. Make the
necessary implementations to be compatible with both versions.