Properly synchronize stopping of macro and its reserved objects
The macro reserved objects e.g. motion or measurement groups are stopped
when the user stops the macro e.g. Ctrl+C in spock, or directly door's
StopMacro command. This stop is executed in a worker thread which is
comletly desynchronized with the macro execution thread.

Fix it forcing the MacroExecutor to wait for the stopping thread to finish its job
before executing the on_stop or on_abort methods. Also protect macros that
catch the StopException by waiting in the mAPI decorator.

Fixes #8.