Fix bug with pseudo counters not involved in acquisition
There is a bug in the acquisition process that affects pseudo counters not
included in the measurement group. The first time the acquisition takes place
this pseudo counters are notified about the value buffer events of the
physical element and their value buffers are filled with the calculation
results. The problem appears in the subsequent acquisitions cause the
pseudo counters buffers are not cleared.

Fix it by clearing the pseudo counters buffers at the beginning of each
acquisition.