Change SCAN_RECORDER_MAP behavior
If the SCAN_RECORDER_MAP is defined it will be the only place where the getRecorderMetaClasses method looks for the recorders. In case the user wants to override just a
single extension and use the rest of the extension as populated in the standard map it requires from him redefining all the map in the SCAN_RECORDER_MAP.

Change this behavior and if the extension is not defined in the
CUSTOM_SCAN_RECORDER_MAP then the recorder will be looked in the standard
map.