Report error if macro parameter definition is wrong
Report error in spock if macro could not be loaded
using relmaclib because of a wrong parameter definition
exists in the macro code.

In this case, the macro will not be available, and
the macro library shall be reloaded using
addmaclib after correcting the error in the parameter
definition of the affected macro.