Fix error with case sensitivity
Sequencer allows to load from xml and from txt, macros that
contains syntax errors related with the case sensitivity. The
macros are loaded into the sequencer, but they cannot be executed,
because macros are case sensitive.

Do not allow to load macros with syntax errors related with
the case sensitivity. Pop up a message if the syntax is not
correct.