Add optional extrapolation of data in continuous acquisition
The very fist points in the continuous acquisition e.g. ascanct may be
omitted, for example, if the motor updates to the software synchronizer were
not precise/fast enough. In this case the zero order interpolation does not
help if we want to get rid of the NaN values.

Introduce a new concept - extrapolation to help avoiding NaN in the very
first points of measurements. Also add ApplyExtrapolation environment
variable to control this feature.