Start hardware TriggerGates as last ones
Software synchronizer is started as the last one in the synchronization
action when the hardware TriggerGate elements are already working. This is
not a problem if the hardware TriggetGate elements work in the position
domain but if they work in the time domain they may start triggering too
fast with respect to the motor's move - the motors are started to move
by the GScan when the MeasurementGroup Start has finsihed. Make the hardware
TriggerGate elements start as the last ones in the synchronization action
what aslo means the last ones in the MeasurementGroup Start command. This
change does not affect the software synchronization which works in the
position domain.