Improve exception handling in CTScan
Currently StopException is not raised again to the higher layers. Due to that
the GScan.step_scan method is not able to determine if the scan was stopped
by the user or not. Furthermore if the scan was run inside of another macro
the StopException is not propagated to the wrapper macro. Change it and
re-raise the StopException

Currently after ScanException the on_waypoints_end method is not called. This
method does the necessary cleanup. Call it.