Implement timeout/watchdog in continuous scans
Hardware problems (missed triggers or positioning problems) could leave the
acquisition waiting infinitely. Add a timeout/watchdog mechanism that will
stop the measurement group if the estimated acquisition time elapse and
the measurement group does not finish. Add an extra time (15 s) to be
sure that the acquisition will not finish. Fix #136.