Filter out events in pseudo motor
Pseudo motor is subscribed to motor’s events with the `on_element_changed`
callback and it processes all the dial position events propagated via
the following chain: dial position -> position -> motor -> pseudo motor.
It is not necessary, cause the only thing does in the callback is to recalculate
the state and status. State and status information are already well
calculated because at the beginning and at the end of the motion these
are notified with the corresponding events. What is even worse is that
the `on_element_changed` callback is called twice because there is
another event propagation route: dial position -> motor -> pseudo motor.
Filter out events different than state and status for pseudo motor.