Filter out events in motor group
Motor group is subscribed to motor’s events with the `on_element_changed`
callback and it processes all the dial position events propagated via
the following chain: dial position -> position -> motor -> motor group.
It is not necessary, cause the only thing does in the callback is to recalculate
the state and status. State and status information are already well
calculated because at the beginning and at the end of the motion these
are notified with the corresponding events. What is even worse is that
the `on_element_changed` callback is called twice because there is
another event propagation route: dial position -> motor -> motor group.
Filter out events different than state for a motor group.