Reduce state propagation level for motion
Some state events are propagated with too high priority 2, meaning that
they will be not filtered. This is not necessary, cause their change
is obvious - start and stop of the motion implies change of the state.
Reduce it to 1 meaning filtering out states that did not changed. This is
necessary for propagating state events to group elements e.g. pseudo motors
or motor groups.