Adapt continuous scans to FQDN in Taurus 4 model names
taurus-org/taurus#488 introduced FQDN in model names. The channel's
column description is composed from the full_name that is composed in
the Pool and transferred with the measurement group configuration and it
does not contain the FQDN. Adapt CAcquisition.buffer_changed callback
to be compatible both with Taurus3 and Taurus4.