Fix dummy counter value after abort
Dummy counters involved in an acquisition that was aborted always
returns the value corresponing to the integration time even
if the abort happend earlier. Fix it and always return the real value
of the counter. This has a side effect on the noramlly finished
acquisition which from now on does not show the round number.