Avoid "already in motion" error
In Sardana there exists two levels of protection from abusive
actions: Tango state machine and the Sardana operation context.
As it is now, the state changes MOVING -> ON before the operation
context is released, allowing the user to move the motor again.
Very fast execution of successive movements may reach a point
where the state has already changed to ON but the operation context
is still locked. Making the requested motion to be rejected.

The proposed change releases the operation context before setting
a new state.