Advance protection when master is disabled
PR#648 makes the selection of master timer/monitor more relaxed.
If the first timer/monitor is disabled it looks for next ones and tries to
select them for the master timer/monitor. If all of them are disabled we
need to advance the protection in order to raise a self explanatory error.