Revert change for hardware acquisition
Intermediate exceptions in the read methods during the acquisition action
does not precisely mean that the acquisition is failed. This decision is
based on the channel's state. A situation may occur when one of the
intermediate read calls raises an exception, but then the subsequent calls
correctly retrieve the data. The purpose of the put_value call is to
indicate the error via the Value attribute.