Always append value even if it is an error
Always append the value (regardless if it is an error or not an error.
This way, in case of the error the error message is transferred to the
clients, as it was done previously with the put_value call, but this
way we also fill the buffer with a placeholder - SardanaValue with the
error information. The new buffer chunk is then correctly pushed to the
client so they don't wait endlessly for the values which won't arrive.