Abort server startup when it will not exist
Server startup procedure, in case the server is not defined in the DB, asks
the user if he would like to create it. In the case when the user does not
wont it the startup procedure continues e.g. it still creates the directory
for the server logs. Abort the procedure so it won't do anything more.