Avoid errors when acquiring with offline controller (#681)
* Avoid errors when acquiring with offline controller

An experimental channels from an offline controller does not export its
attributes. Then, it is not possible to obtain the Value attribute
configuration and an empty AttributeInfoEx is created for it. Its default
data_type is set to DevVoid. Handle this case and consider the
TangoChannelInfo with an unknown data_type. This avoids errors.