Docs: change home page address to www.sardana-controls.org
Due to unknown reasons the sardana-controls.org address does not work
correctly with RTD. Change the Home Page address to a working
address www.sardana-controls.org. In the near future, if we finally
follow the steps of Taurus project to move the docs to github pages,
the sardana-controls.org address should work again.

The same problem applies to sardana-scada.org