Fix test when TANGO_HOST with PQDN is used (Taurus 3)
DeviceProxy.get_db_host always return FQDN even if the database was
configured with PQDN via the TANGO_HOST environment variable.

The Sardana servers compose the device full names taking into account if
the database was configured with the PQDN via the TANGO_HOST environmnet
variable, and in this case use it without expanding it to the FQDN.
Do the same in the test in order to avoid problesm with full name matching.