Use a dedicated pool with only one thread for ValueBuffer events
The global pool of threads in the MacroServer contains 10 threads.
CAcquisition adds jobs for handling the ValueBuffer events to this pool.
Serialized execution of this jobs is assumed by the downstream logic
(composing of completed records).

The critical section protection (RLock) in the in the RecordList.addData
method does not guarantee the serialized execution of jobs.
This is because multiple jobs may be deposited on the queue and multiple
free threads may take them. Even if they will wait for the lock the
thread that will proceed when the lock is released is undefined and may
vary across implementations.