Add configuration attribute to DiscretePseudoMotorController
Provide a more pythonic interface to the discrete pseudo motor
configuration.

The new configuration attribute encapsulates the values
from the label and calibration attributes as a dictionary.

The dictionary keys are extracted from the labels attribute.
and each dictionary entry is also a dictionary with four keys:
a `pos` key corresponding to the label enumerated position
integer value and three float values corresponding to the
calibration values and named as `min`, `set`, and `max`.