Implement fallback for getPoolObject in case it was not filled
Override BaseElement.getPoolObj because the reference to pool object may
not be filled. This reference is filled when the element is obtained
using Pool.getObject. If one obtain the element directly using Taurus
e.g. mot = taurus.Device(<mot_name>) it won't be filled. In this case
look for the pool object using the database information.