Fix snapshots list loading in the expconf widget.
This commit fixes the read of PreScanSnapshots elements. The expconf
widget was loading the snapshot list from a file, without checking if it was
FQDN compatible or not, therefore, it was always appending a "tango://"
prefix to all the elements. Now is checking if this appending is needed
or not.