Fix pseudo counters in continuous acquisition.
Let's consider the folowing example. A pseudo coutner is based on two
counters. If counter#1 reported value of index 1 and 3 in separate
chunks and then counter#2 reported values 1, 2 and 3 in the same chunk,
the pseudo counter won't calculate the index 3. This is because the
LateValueException is returning from the on_change callback, ignoring
all the following indexes. Fix it by continuing the loop and just
skipping calculation for late values.