Move pre-scan and post-scan hooks out of scan_loop
pre-scan and post-scan hooks are executed within the scan_loop. Before
the scan loop, for example, are started the scan recorders making
them to start the scan entry. It makes more sense to not start them
until the pre-scan hooks are executed - they may want discard the scan
or simply may not want to mix the scan header output with the pre-scan
hooks output.

This solution allows to avoid the code duplication in all the scan
specific implementation.