Adapt to new signal style
Adapt to new signal style.

The signal "modelChanged()" is only used to call
the method onModelChanged. No other widgets are using this
signal, and it is not easy to adapt it to the new PyQt
signal style without finding collisions with a Taurus signal
with the same name, etc.

Applied solution:
After discussion, decided to remove signal "modelChanged()"
and its emits, and call directly the method onModelChanged instead.