Use keyword args in mixin classes
Use keyword args instead of positional args in mixin classes.

PyQt5 and PyQt4 differ in the way they treat the multiple
inheritance.
This is causing exceptions when instantiating classes that inherit
from a Qt class and a mixin class.
A workaround seems to be to use only kwargs in the init of the
mixins.

TODO: This requires more attention to check if everything is really
working as it should.