Fix bad syntax in DoorAttrListener
DoorAttrListener is a deprecated and unused class that wil soon be
removed. Its old code (using old-style signals) emitted a signal called
door<DOOR_NAME>Changed. An attempt to port it to new-style signals was
done in a previous copmmit, but it is not ok. remove the signal emit
altogether (not worth wasting time in this class that will be removed).