Change default timer to __default and refactor code
Setting default timer to __default required code refactoring.
Now part of the configuration is created at the channel's construction
time and other part (the one related to the timer) is created on demand
when acquiring. This is because the default timer may not exist yet at the
channel's construction time e.g. when Tango devices are initialized.