Separate GH tests into two classes
Some general hook tests refer just to the configuration macros
and some others also execute the macro to which the general hook
was attached in order to check if it could be executed.

Split the tests in two groups cause the setUp and tearDown should
be different. This avoids problems of deletion of the measurement
group when it was never created.