Get ms_properties from db properties
Get ms_properties from db properties; only in case it is None,
build the path to ms_properties and normalize it to be adapted
to the given OS  (linux / windows).

Try to remove the file, and if it is not possible, give an
information message (removing the file macroserver.properties,
sometimes is giving problems in Windows).