Adapt to removal of "taurus popup menu" feature
The PR https://github.com/taurus-org/taurus/pull/906 in taurus disables
the implicit showing of self.taurusMenu as a context menu.

This affects to MacroExecutionWindow (used by macroexecutor and
macrosequncer) since it relies on this feature to show its context menu.

Explicitly overwrite contextMenuEvent to provide the expected behaviour
in this case.