Add auto-deploy to PyPI with Travis
Make Travis deploy to PyPI when a version tag (major.minor.patch) is
pushed. It uses an ad-hoc PyPI account (sardana_bot) which was given
Maintainer permisions in PyPI. The sardana_bot password is encrypted
(using the travis-encrypt python module) with the sardana-org/sardana
Travis public key so that Travis can decrypt it for deployment.