Don't change the source information in meas grp conf
source information (Value) is used to determine data type and shape
(from max dims) experimental channel from the Value attribute readout in
expconf. Until we provide a better mechanism for that it is better
to keep the source always Value instead of alternating it with ValueRef if
this one is enabled. Ideally this should be like this...