Add py2_round to sardana utils
The behavior of round has changed in Python 3. In Python 2, rounding of halfway cases was away from zero,
and round() would always return a float. In Python 3 rounding of halfway
cases are now always towards the nearest even. This is standard practice,
as it will make a set of evenly distributed roundings average out.