py3: don't import when defining class
Sardana import was delayed until class definition, most probably this still,
comes from the times when the sardana-taurus widgets were in the Taurus project.
Then it had sense cause in Taurus we did not want to have sardana dependency.
Since these widgets are now part of Sardana project the import can be advanced.

The reason of changing this now is that this way of importing has some problems on
Python 3.